refactor(string-calculator) : NumbersFactory로 숫자배열 추출 객체 분리.
